streamAdapter-base/src/test/java/com/ericsson/streamAdapter/util/EventMetricsTest.java
/*------------------------------------------------------------------------------
 *******************************************************************************
 * COPYRIGHT Ericsson 2013
 *
 * The copyright to the computer program(s) herein is the property of
 * Ericsson Inc. The programs may be used and/or copied only with written
 * permission from Ericsson Inc. or in accordance with the terms and
 * conditions stipulated in the agreement/contract under which the
 * program(s) have been supplied.
 *******************************************************************************
 *----------------------------------------------------------------------------*/
package com.ericsson.streamAdapter.util;

import static org.junit.Assert.*;

import org.junit.Test;

public class EventMetricsTest {

	@Test 
	public void test1() {
		EventMetrics em = new EventMetrics(false);
		em.incrementInits();
		em.incrementConnects();
		em.incrementDisconnects();
		em.incrementInvalidRecords(1);
		em.incrementEvents();
		em.incrementRecords();
		String res = em.toString();
        System.out.println(res);
		// There is a check that the record counts match, which resets the record counts
//		assertEquals("Unexpected output", res,
//				" Number of Events:1 Inits:1 Connects:1 Disconnects:1 No sources:0 Total:[Per 0.0sec interval]1 per sec:0");
//		res = em.toString();
//		assertEquals("Unexpected output", res,
//				" Number of Events:1 Inits:1 Connects:1 Disconnects:1 No sources:0 Total:[Per 0.0sec interval]1 per sec:0");
        assertTrue(true);
	}

	@Test 
	public void test2() {
//		EventMetrics em = new EventMetrics(false);
//		em.incrementInvalidRecords(2);
//		em.incrementInits();
//		em.incrementConnects();
//		em.incrementConnects();
//		em.incrementDisconnects();
//		em.incrementDisconnects();
//		em.incrementDisconnects();
//		em.incrementEvents();
//		em.incrementEvents();
//		em.incrementEvents();
//		em.incrementEvents();
//		em.incrementRecords();
//		em.incrementRecords();
//		em.incrementRecords();
//		em.incrementRecords();
//		em.incrementRecords();
//		em.incrementRecords();
//		em.incrementRecords();
//		em.incrementRecords();
//		String res = em.toString();
//		assertEquals("Unexpected output", res,
//				" Number of Events:4 Inits:1 Connects:2 Disconnects:3 No sources:0 Total:8 per sec:0");
//		res = em.toString();
//		// There is a check that the record counts match, which resets the record counts
//		assertEquals("Unexpected output", res,
//				" Number of Events:4 Inits:1 Connects:2 Disconnects:3 No sources:0 Total:12 per sec:0");
        assertTrue(true);
	}

	@Test 
	public void test3() {
		// test aggregated counters
//		EventMetrics em = new EventMetrics(true);
//		em.incrementInits();
//		em.incrementConnects();
//		em.incrementDisconnects();
//		em.incrementInvalidRecords(1);
//		em.incrementRecords();
//		String res = em.toString();
//		// There is a check that the record counts match, which resets the record counts
////		assertEquals("Unexpected output", res,
////				" Total:1 per sec:0");
//		res = em.toString();
//		em.incrementRecords();
//		em.incrementRecords();
//		res = em.toString();
////		assertEquals("Unexpected output", res,
////				" Total:3 per sec:0");
        assertTrue(true);
	}

}
